Fix the R2R
When EOS happenes, the PD will output the frames to following kernels,
which will cause some timing issue.
For example, PD will output (48, 40 ….) only when the first minigop
(poc32,24,…31) is done.
But when EOS happened, PD will output (POC64,56..) when previous minigop
is not finished yet.
For example: POC44 is still on-going, and TPL56 will use POC44,
sometimes recon of 44 is generated, sometimes not, and will cause r2r.
M6 doesn’t happen probably because we only use single ref for M6. 56=>44
is 3rd ref in L1.

The fix/workaround is simple, by limiting the size of min_me.
Now we use
        min_me = mg_size + 1 + 6; //TODO add Constant for 6
But since we don’t use the trailing frames right now, so we should change it to
        min_me = mg_size + 1;

Signed-off-by: Jing Li <jing.b.li@intel.com>